Implemented addToIndex and getNextUrl.
Changed the interface of Spider - replaced method getText with getKeyWords (all the words of the document, every word is there only 1 time). It will return set of keywords in the document. Will be much faster to search in - in linear time we will not go through the whole document but just through the keywords.